app/requirements.txt
flask==3.0.2
google-cloud-firestore==2.15.0
google-cloud-storage==2.14.0
python-dotenv==1.0.1
gunicorn==21.2.0
pytest==8.0.0
pytest-cov==6.1.1 